Added maven to build the project and updated unit tests for Sorting Algorithms accordingly
