+++ coollime/src/main/java/coollime/App.java
@@ -1,0 +1,13 @@
+package coollime;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
