+++ coollime/src/test/java/coollime/AppTest.java
@@ -1,0 +1,18 @@
+package coollime;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest {
+  /**
+   * Rigorous Test :-)
+   */
+  @Test
+  public void shouldAnswerWithTrue() {
+    assertTrue(true);
+  }
+}
